feat: basic login with spotify
